+++ instrumentation-genai/opentelemetry-instrumentation-google-genai/tests/generate_content/nonstreaming_base.py
@@ -260,12 +260,26 @@ class NonStreamingTestCase(TestCase):
         )
 
     def test_generated_span_counts_tokens(self):
-        self.configure_valid_response(input_tokens=123, output_tokens=456)
+        self.configure_valid_response(
+            input_tokens=123,
+            output_tokens=456,
+            cached_tokens=50,
+            thinking_tokens=17,
+        )
         self.generate_content(model="gemini-2.0-flash", contents="Some input")
         self.otel.assert_has_span_named("generate_content gemini-2.0-flash")
         span = self.otel.get_span_named("generate_content gemini-2.0-flash")
         self.assertEqual(span.attributes["gen_ai.usage.input_tokens"], 123)
-        self.assertEqual(span.attributes["gen_ai.usage.output_tokens"], 456)
+        self.assertEqual(
+            span.attributes["gen_ai.usage.output_tokens"], 456 + 17
+        )
+        # New sem conv should not appear when flag is not experimental mode..
+        self.assertNotIn(
+            "gen_ai.usage.cache_read.input_tokens", span.attributes
+        )
+        self.assertNotIn(
+            "gen_ai.usage.reasoning.output_tokens", span.attributes
+        )
 
     @patch.dict(
         "os.environ",
@@ -445,7 +459,11 @@ class NonStreamingTestCase(TestCase):
             ):
                 self.setUp()
                 with patched_environ, patched_otel_mapping:
-                    self.configure_valid_response(text=output)
+                    self.configure_valid_response(
+                        text=output,
+                        cached_tokens=50,
+                        thinking_tokens=17,
+                    )
                     self.generate_content(
                         model="gemini-2.0-flash",
                         contents=content,
@@ -460,6 +478,22 @@ class NonStreamingTestCase(TestCase):
                     )
                     event = self.otel.get_event_named(
                         "gen_ai.client.inference.operation.details"
+                    )
+                    self.assertEqual(
+                        event.attributes[
+                            "gen_ai.usage.cache_read.input_tokens"
+                        ],
+                        50,
+                    )
+                    self.assertEqual(
+                        event.attributes[
+                            "gen_ai.usage.reasoning.output_tokens"
+                        ],
+                        17,
+                    )
+                    self.assertEqual(
+                        event.attributes["gen_ai.usage.output_tokens"],
+                        17,
                     )
                     assert (
                         event.attributes[
@@ -765,7 +799,11 @@ class NonStreamingTestCase(TestCase):
             ):
                 self.setUp()
                 with patched_environ, patched_otel_mapping:
-                    self.configure_valid_response(text="Some response content")
+                    self.configure_valid_response(
+                        text="Some response content",
+                        cached_tokens=50,
+                        thinking_tokens=19,
+                    )
                     self.generate_content(
                         model="gemini-2.0-flash",
                         contents="Some input",
@@ -777,6 +815,22 @@ class NonStreamingTestCase(TestCase):
                     )
                     span = self.otel.get_span_named(
                         "generate_content gemini-2.0-flash"
+                    )
+                    self.assertEqual(
+                        span.attributes[
+                            "gen_ai.usage.cache_read.input_tokens"
+                        ],
+                        50,
+                    )
+                    self.assertEqual(
+                        span.attributes[
+                            "gen_ai.usage.reasoning.output_tokens"
+                        ],
+                        19,
+                    )
+                    self.assertEqual(
+                        span.attributes["gen_ai.usage.output_tokens"],
+                        19,
                     )
                     if mode in [
                         ContentCapturingMode.SPAN_ONLY,