+++ instrumentation-genai/opentelemetry-instrumentation-google-genai/tests/generate_content/util.py
@@ -25,7 +25,9 @@ def create_response(
     candidates: Optional[list[genai_types.Candidate]] = None,
     text: Optional[str] = None,
     input_tokens: Optional[int] = None,
+    thinking_tokens: Optional[int] = None,
     output_tokens: Optional[int] = None,
+    cached_tokens: Optional[int] = None,
     model_version: Optional[str] = None,
     usage_metadata: Optional[
         genai_types.GenerateContentResponseUsageMetadata
@@ -53,6 +55,10 @@ def create_response(
         usage_metadata.prompt_token_count = input_tokens
     if output_tokens is not None:
         usage_metadata.candidates_token_count = output_tokens
+    if cached_tokens is not None:
+        usage_metadata.cached_content_token_count = cached_tokens
+    if thinking_tokens is not None:
+        usage_metadata.thoughts_token_count = thinking_tokens
     return genai_types.GenerateContentResponse(
         candidates=candidates,
         usage_metadata=usage_metadata,